+++ packages/dive/lib/src/dive_audio_meter_source.dart
@@ -113,7 +113,6 @@ class DiveAudioMeterStateNotifier extends StateNotifier<DiveAudioMeterState> {
 
   void updateState(DiveAudioMeterState stateModel) {
     state = stateModel;
-    print("DiveAudioMeterState updated: $stateModel");
   }
 }